unk8sify many auth tests
